+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Era.hs
@@ -8,14 +8,21 @@
 module Test.Cardano.Ledger.Alonzo.Era (
   module Test.Cardano.Ledger.Mary.Era,
   AlonzoEraTest,
+  mkTestLedgerTxInfo,
 ) where
 
 import Cardano.Ledger.Alonzo
 import Cardano.Ledger.Alonzo.Core
 import Cardano.Ledger.Alonzo.Plutus.Context
 import Cardano.Ledger.Alonzo.UTxO
+import Cardano.Ledger.BaseTypes (ProtVer)
 import Cardano.Ledger.Plutus (Language (..))
+import Cardano.Ledger.State
+import Cardano.Slotting.EpochInfo (EpochInfo)
+import Cardano.Slotting.Time (SystemStart)
+import Data.Text (Text)
 import Data.TreeDiff
+import Lens.Micro
 import Paths_cardano_ledger_alonzo (getDataFileName)
 import Test.Cardano.Ledger.Alonzo.Arbitrary ()
 import Test.Cardano.Ledger.Alonzo.Binary.Annotator ()
@@ -81,3 +88,29 @@ instance AllegraEraTest AlonzoEra
 instance MaryEraTest AlonzoEra
 
 instance AlonzoEraTest AlonzoEra
+
+-- | This is a construction of `LedgerTxInfo` without any memoization.
+mkTestLedgerTxInfo ::
+  (EraUTxO era, EraPlutusContext era, ScriptsNeeded era ~ AlonzoScriptsNeeded era) =>
+  ProtVer ->
+  EpochInfo (Either Text) ->
+  SystemStart ->
+  UTxO era ->
+  Tx level era ->
+  LedgerTxInfo era
+mkTestLedgerTxInfo protVer epochInfo systemStart utxo tx =
+  let
+    scriptsProvided = getScriptsProvided utxo tx
+    scriptsNeeded = getScriptsNeeded utxo (tx ^. bodyTxL)
+    (_, plutusScriptsUsed) =
+      resolveNeededPlutusScriptsWithPurpose protVer scriptsProvided scriptsNeeded mempty
+   in
+    LedgerTxInfo
+      { ltiProtVer = protVer
+      , ltiEpochInfo = epochInfo
+      , ltiSystemStart = systemStart
+      , ltiUTxO = utxo
+      , ltiTx = tx
+      , ltiScriptsUsed = plutusScriptsUsed
+      , ltiMemoizedSubTransactions = mempty
+      }